FINISHED/PIXTRAL/jax/pyproject.toml
[project]
name = "jax_pixtral"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "einops>=0.8.1",
    "jax[cuda12]>=0.6.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "safetensors>=0.5.3",
]

